+++ commandline/src/test/java/tech/pegasys/web3signer/commandline/valueprovider/DemoCommand.java
@@ -32,9 +32,19 @@ public class DemoCommand {
   @Option(names = "--name", description = "Name")
   String name;
 
+  @Option(
+      names = {"--alias", "--aliases"},
+      description = "Aliases")
+  String alias;
+
   @Command(name = "country", description = "Country Codes")
   static class SubCommand {
     @Option(names = "--codes", split = ",")
     List<String> countryCodes;
+
+    @Option(
+        names = {"--subalias", "--subaliases"},
+        description = "Subcommand Alias")
+    String subalias;
   }
 }