+++ commandline/src/test/java/tech/pegasys/web3signer/commandline/valueprovider/EnvironmentVariableDefaultProviderTest.java
@@ -83,6 +83,29 @@ class EnvironmentVariableDefaultProviderTest {
     assertThat(subCommand.countryCodes).isNullOrEmpty();
   }
 
+  @Test
+  void environmentVariablesAreCreatedForAliases() {
+
+    assertAliasCreated("DEMO_ALIAS", "testValue");
+    assertAliasCreated("DEMO_ALIASES", "testValue2");
+    assertAliasCreated("DEMO_COUNTRY_SUBALIAS", "testValue3", "country");
+    assertAliasCreated("DEMO_COUNTRY_SUBALIASES", "testValue4", "country");
+  }
+
+  private void assertAliasCreated(String envKey, String envValue, String... subCommandName) {
+    final EnvironmentVariableDefaultProvider defaultProvider =
+        new EnvironmentVariableDefaultProvider(Map.of(envKey, envValue));
+
+    commandLine.setDefaultValueProvider(defaultProvider);
+    commandLine.parseArgs(subCommandName);
+
+    if (subCommandName.length > 0) {
+      assertThat(subCommand.subalias).isEqualTo(envValue);
+    } else {
+      assertThat(demoCommand.alias).isEqualTo(envValue);
+    }
+  }
+
   private Map<String, String> validEnvMap() {
     return Map.of(
         "DEMO_X", "10", "DEMO_Y", "20", "DEMO_NAME", "test name", "DEMO_COUNTRY_CODES", "AU,US");